backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
pydantic==1.8.2
python-dotenv==0.19.0
sqlalchemy==1.4.23
bcrypt
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
requests==2.31.0
